+++ libre_pvz_resources/src/dynamic.rs
@@ -18,9 +18,10 @@
 
 //! Dynamic resource support. For simplicity, we use shared global state for registry.
 
-use std::any::{Any, TypeId};
+use std::any::TypeId;
 use std::collections::BTreeMap;
 use std::fmt::{Debug, Formatter};
+use std::ops::Deref;
 use bevy::prelude::*;
 use bevy::reflect::erased_serde::{
     Serialize as ErasedSerialize,
@@ -74,6 +75,9 @@ impl DynamicRegistry {
         GLOBAL_REGISTRY.get().expect("DynamicRegistry: must initialize before use")
     }
 
+    /// Get a shared reference to Bevy's type registry (already wrapped in `Arc`).
+    pub fn get_bevy_type_registry(&self) -> &TypeRegistry { &self.type_registry }
+
     /// Get the [`ReflectAnyResource`] for the type registered as the given name.
     pub fn resource_by_name(&self, name: &str) -> Option<ReflectAnyResource> {
         let id = self.readable_name_to_id.read().get(name).copied()?;
@@ -105,11 +109,11 @@ pub trait AnyResource: Reflect + ErasedSerialize + Send + Sync + 'static {
     /// Convert to a mutable [`Reflect`] trait object.
     fn as_reflect_mut(&mut self) -> &mut dyn Reflect;
     /// Deserialize from an [`erased`](bevy::reflect::erased_serde) deserializer.
-    fn erased_deserialize(src: &mut dyn ErasedDeserializer) -> Result<DynamicResource, Error> where Self: Sized;
+    fn erased_deserialize(src: &mut dyn ErasedDeserializer) -> Result<Box<dyn AnyResource>, Error> where Self: Sized;
     /// Encode as [`bincode`] to a given [`Writer`] using standard configuration.
     fn erased_encode(&self, writer: &mut dyn Writer) -> Result<(), EncodeError>;
     /// Decode as [`bincode`] from a given [`Reader`] using standard configuration.
-    fn erased_decode(reader: &mut dyn Reader) -> Result<DynamicResource, DecodeError> where Self: Sized;
+    fn erased_decode(reader: &mut dyn Reader) -> Result<Box<dyn AnyResource>, DecodeError> where Self: Sized;
 }
 
 /// [`TypeData`](bevy::reflect::TypeData) providing support for [`AnyResource`] trait.
@@ -119,9 +123,9 @@ pub trait AnyResource: Reflect + ErasedSerialize + Send + Sync + 'static {
 pub struct ReflectAnyResource {
     get: fn(&dyn Reflect) -> &dyn AnyResource,
     get_mut: fn(&mut dyn Reflect) -> &mut dyn AnyResource,
-    get_boxed: fn(Box<dyn Reflect>) -> DynamicResource,
-    erased_deserialize: fn(&mut dyn ErasedDeserializer) -> Result<DynamicResource, Error>,
-    erased_decode: fn(&mut dyn Reader) -> Result<DynamicResource, DecodeError>,
+    get_boxed: fn(Box<dyn Reflect>) -> Box<dyn AnyResource>,
+    erased_deserialize: fn(&mut dyn ErasedDeserializer) -> Result<Box<dyn AnyResource>, Error>,
+    erased_decode: fn(&mut dyn Reader) -> Result<Box<dyn AnyResource>, DecodeError>,
 }
 
 impl ReflectAnyResource {
@@ -129,14 +133,14 @@ impl ReflectAnyResource {
     pub fn get<'a>(&self, val: &'a dyn Reflect) -> &'a dyn AnyResource { (self.get)(val) }
     /// Try to downcast a `&mut dyn Reflect` type to `&mut dyn AnyResource`.
     pub fn get_mut<'a>(&self, val: &'a mut dyn Reflect) -> &'a mut dyn AnyResource { (self.get_mut)(val) }
-    /// Try to downcast a `Box<dyn Reflect>` type to `DynamicResource`.
-    pub fn get_boxed(&self, val: Box<dyn Reflect>) -> DynamicResource { (self.get_boxed)(val) }
+    /// Try to downcast a `Box<dyn Reflect>` type to `Box<dyn AnyResource>`.
+    pub fn get_boxed(&self, val: Box<dyn Reflect>) -> Box<dyn AnyResource> { (self.get_boxed)(val) }
     /// Deserialize using [`serde`] into a trait object.
-    pub fn erased_deserialize(&self, src: &mut dyn ErasedDeserializer) -> Result<DynamicResource, Error> {
+    pub fn erased_deserialize(&self, src: &mut dyn ErasedDeserializer) -> Result<Box<dyn AnyResource>, Error> {
         (self.erased_deserialize)(src)
     }
     /// Deserialize using [`bincode`] into a trait object.
-    pub fn erased_decode(&self, reader: &mut dyn Reader) -> Result<DynamicResource, DecodeError> {
+    pub fn erased_decode(&self, reader: &mut dyn Reader) -> Result<Box<dyn AnyResource>, DecodeError> {
         (self.erased_decode)(reader)
     }
 }
@@ -164,14 +168,14 @@ impl<T> AnyResource for T
     fn into_reflect(self: Box<Self>) -> Box<dyn Reflect> { self }
     fn as_reflect(&self) -> &dyn Reflect { self }
     fn as_reflect_mut(&mut self) -> &mut dyn Reflect { self }
-    fn erased_deserialize(src: &mut dyn ErasedDeserializer) -> Result<DynamicResource, Error> where Self: Sized {
+    fn erased_deserialize(src: &mut dyn ErasedDeserializer) -> Result<Box<dyn AnyResource>, Error> where Self: Sized {
         T::deserialize(src).map(|x| Box::new(x) as _)
     }
     fn erased_encode(&self, writer: &mut dyn Writer) -> Result<(), EncodeError> {
         // hopefully this gets inlined, and the double reference is optimised away
         T::encode(self, &mut EncoderImpl::new(Proxy(writer), BINCODE_CONFIG))
     }
-    fn erased_decode(reader: &mut dyn Reader) -> Result<DynamicResource, DecodeError> where Self: Sized {
+    fn erased_decode(reader: &mut dyn Reader) -> Result<Box<dyn AnyResource>, DecodeError> where Self: Sized {
         // hopefully this gets inlined, and the double reference is optimised away
         T::decode(&mut DecoderImpl::new(Proxy(reader), BINCODE_CONFIG)).map(|x| Box::new(x) as _)
     }
@@ -181,43 +185,45 @@ impl Debug for dyn AnyResource {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result { self.debug(f) }
 }
 
-/// Resource, but without a statically-known type.
-pub type DynamicResource = Box<dyn AnyResource>;
+struct Wrapper<'a, T: ?Sized>(&'a T);
 
-struct Wrapper<'a>(&'a dyn AnyResource);
-
-impl<'a> Serialize for Wrapper<'a> {
+impl<'a, T: AnyResource + ?Sized> Serialize for Wrapper<'a, T> {
     #[inline(always)]
     fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
         erased_serde_serialize(self, serializer)
     }
 }
 
-impl Serialize for DynamicResource {
+fn serialize_any_resource<T, S>(value: &T, serializer: S) -> Result<S::Ok, S::Error>
+    where T: AnyResource + ?Sized, S: Serializer {
+    let g = DynamicRegistry::global().readable_name_from_id.read();
+    let name = g.get(&value.type_id()).map(Box::as_ref)
+        .ok_or_else(|| S::Error::custom(format_args!(
+            "type '{}' does not support dynamic serialization", value.type_name())))?;
+    let mut map = serializer.serialize_map(Some(2))?;
+    map.serialize_entry(name, &Wrapper(value))?;
+    map.end()
+}
+
+impl Serialize for dyn AnyResource {
     fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        let g = DynamicRegistry::global().readable_name_from_id.read();
-        let name = g.get(&self.type_id()).map(Box::as_ref)
-            .ok_or_else(|| S::Error::custom(format_args!(
-                "type '{}' does not support dynamic serialization", self.type_name())))?;
-        let mut map = serializer.serialize_map(Some(2))?;
-        map.serialize_entry(name, &Wrapper(self.as_ref()))?;
-        map.end()
+        serialize_any_resource(self, serializer)
     }
 }
 
 impl<'de> DeserializeSeed<'de> for ReflectAnyResource {
-    type Value = DynamicResource;
+    type Value = Box<dyn AnyResource>;
     fn deserialize<D: Deserializer<'de>>(self, deserializer: D) -> Result<Self::Value, D::Error> {
         self.erased_deserialize(&mut <dyn ErasedDeserializer>::erase(deserializer))
             .map_err(D::Error::custom)
     }
 }
 
-impl<'de> Deserialize<'de> for DynamicResource {
+impl<'de> Deserialize<'de> for Box<dyn AnyResource> {
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
         struct DynResVisitor;
         impl<'de> Visitor<'de> for DynResVisitor {
-            type Value = DynamicResource;
+            type Value = Box<dyn AnyResource>;
             fn expecting(&self, formatter: &mut Formatter) -> std::fmt::Result {
                 formatter.write_str("a dynamic resource")
             }
@@ -237,18 +243,29 @@ impl<'de> Deserialize<'de> for DynamicResource {
     }
 }
 
-impl Encode for DynamicResource {
+fn encode_any_resource<T, E>(value: &T, encoder: &mut E) -> Result<(), EncodeError>
+    where T: AnyResource + ?Sized, E: Encoder {
+    let g = DynamicRegistry::global().readable_name_from_id.read();
+    let name = g.get(&value.type_id()).map(Box::as_ref)
+        .ok_or_else(|| EncodeError::OtherString(format!(
+            "type '{}' does not support dynamic serialization", value.type_name())))?;
+    name.encode(encoder)?;
+    value.erased_encode(encoder.writer())
+}
+
+impl Encode for dyn AnyResource {
     fn encode<E: Encoder>(&self, encoder: &mut E) -> Result<(), EncodeError> {
-        let g = DynamicRegistry::global().readable_name_from_id.read();
-        let name = g.get(&self.type_id()).map(Box::as_ref)
-            .ok_or_else(|| EncodeError::OtherString(format!(
-                "type '{}' does not support dynamic serialization", self.type_name())))?;
-        name.encode(encoder)?;
-        self.erased_encode(encoder.writer())
+        encode_any_resource(self, encoder)
     }
 }
 
-impl Decode for DynamicResource {
+impl Encode for Box<dyn AnyResource> {
+    fn encode<E: Encoder>(&self, encoder: &mut E) -> Result<(), EncodeError> {
+        encode_any_resource(self.as_ref(), encoder)
+    }
+}
+
+impl Decode for Box<dyn AnyResource> {
     fn decode<D: Decoder>(decoder: &mut D) -> Result<Self, DecodeError> {
         let name = String::decode(decoder)?;
         let reg = DynamicRegistry::global();
@@ -257,5 +274,96 @@ impl Decode for DynamicResource {
                 format!("type {} not registered for dynamic deserialization", name)
             ))?;
         reflect.erased_decode(decoder.reader())
+    }
+}
+
+/// Resource, but without a statically-known type.
+#[repr(transparent)]
+pub struct DynamicResource<T: ?Sized>(Box<T>);
+
+impl<T: ?Sized> From<Box<T>> for DynamicResource<T> {
+    fn from(value: Box<T>) -> Self { DynamicResource(value) }
+}
+
+impl<T: ErasedResource + ?Sized> Debug for DynamicResource<T> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        self.debug(f)
+    }
+}
+
+/// Allow dynamic (de)serialization for trait objects.
+pub trait ErasedResource: AnyResource {
+    /// Readable name for this type of trait objects as dynamic resource (for diagnostics).
+    const RESOURCE_TYPE: &'static str;
+    /// Try convert from a fully-erased dynamic resource to this trait object.
+    fn try_from_erased(erased: Box<dyn AnyResource>) -> Result<DynamicResource<Self>, String>;
+}
+
+/// Mark a trait available for dynamic (de)serialization.
+/// ```
+/// # use bevy::reflect::reflect_trait;
+/// # use libre_pvz_resources::dynamic::AnyResource;
+/// # use libre_pvz_resources::mark_trait_as_dynamic_resource;
+/// #[reflect_trait]
+/// pub trait DoThing: AnyResource {
+///     fn do_thing(&self) -> String;
+/// }
+/// mark_trait_as_dynamic_resource!(DoThing, ReflectDoThing);
+/// ```
+#[macro_export]
+macro_rules! mark_trait_as_dynamic_resource {
+    ($trait_name:ident, $reflect_trait:ty $(,)?) => {
+        $crate::mark_trait_as_dynamic_resource! {
+            $trait_name, $reflect_trait, stringify!($trait_name),
+        }
+    };
+    ($trait_name:ident, $reflect_trait:ty, $readable_name:expr $(,)?) => {
+        impl $crate::dynamic::ErasedResource for dyn $trait_name {
+            const RESOURCE_TYPE: &'static str = $readable_name;
+            fn try_from_erased(erased: Box<dyn $crate::dynamic::AnyResource>)
+                    -> Result<$crate::dynamic::DynamicResource<Self>, String> {
+                let reg = $crate::dynamic::DynamicRegistry::global().get_bevy_type_registry().read();
+                let report_error = |real_type: &str| {
+                    let expected = Self::RESOURCE_TYPE;
+                    format!("{real_type} is not an instance of {expected}")
+                };
+                let reflect = reg.get_type_data::<$reflect_trait>(erased.type_id())
+                    .ok_or_else(|| report_error(erased.type_name()))?;
+                reflect.get_boxed(erased.into_reflect())
+                    .map($crate::dynamic::DynamicResource::from)
+                    .map_err(|e| report_error(e.type_name()))
+            }
+        }
+    }
+}
+
+impl<T: ?Sized> Deref for DynamicResource<T> {
+    type Target = T;
+    fn deref(&self) -> &Self::Target { self.0.as_ref() }
+}
+
+impl<T: ErasedResource + ?Sized> Serialize for DynamicResource<T> {
+    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
+        serialize_any_resource(self.0.as_ref(), serializer)
+    }
+}
+
+impl<'de, T: ErasedResource + ?Sized> Deserialize<'de> for DynamicResource<T> {
+    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
+        let erased = Box::<dyn AnyResource>::deserialize(deserializer)?;
+        T::try_from_erased(erased).map_err(D::Error::custom)
+    }
+}
+
+impl<T: ErasedResource + ?Sized> Encode for DynamicResource<T> {
+    fn encode<E: Encoder>(&self, encoder: &mut E) -> Result<(), EncodeError> {
+        encode_any_resource(self.0.as_ref(), encoder)
+    }
+}
+
+impl<T: ErasedResource + ?Sized> Decode for DynamicResource<T> {
+    fn decode<D: Decoder>(decoder: &mut D) -> Result<Self, DecodeError> {
+        let erased = Box::<dyn AnyResource>::decode(decoder)?;
+        T::try_from_erased(erased).map_err(DecodeError::OtherString)
     }
 }